aes: remove CFG_TA_OPTEE_CORE_API_COMPAT_1_1
Remove CFG_TA_OPTEE_CORE_API_COMPAT_1_1=y to fix TA panic 0x0 from
null handle operations.
Change tag_len/out_len from uint32_t to size_t in auth_enc_op() to match
modern TEE API function signatures for TEE_AEEncryptFinal and
TEE_AEDecryptFinal.
Fix format specifiers from %d to %zu for size_t variables to resolve
compiler warnings.

Acked-by: Jens Wiklander <jens.wiklander@linaro.org>

+++ aes/ta/aes_ta.c
@@ -327,7 +327,7 @@ static TEE_Result cipher_buffer(void *session, uint32_t param_types,
 		return TEE_ERROR_BAD_PARAMETERS;
 
 	if (params[1].memref.size < params[0].memref.size) {
-		EMSG("Bad sizes: in %d, out %d", params[0].memref.size,
+		EMSG("Bad sizes: in %zu, out %zu", params[0].memref.size,
 						 params[1].memref.size);
 		return TEE_ERROR_BAD_PARAMETERS;
 	}
@@ -347,8 +347,8 @@ static TEE_Result auth_enc_op(void *session, uint32_t param_types, TEE_Param par
 {
 	struct aes_cipher *sess = session;
 	TEE_Result res = TEE_ERROR_OUT_OF_MEMORY;
-	uint32_t tag_len = 0;
-	uint32_t out_len = 0;
+	size_t tag_len = 0;
+	size_t out_len = 0;
 	void *in_buf = NULL;
 	size_t in_sz = 0;
 	bool encrypt = true;
